refactor: more refactoring
