+++ components/Layout/index.ts
@@ -1,2 +1,4 @@
+export { default as AppRoot } from './AppRoot'
+export { default as Page } from './Page'
 export { default as PageBody } from './PageBody'
 export { default as PageHeader } from './PageHeader'
