+++ pages/about.tsx
@@ -2,9 +2,7 @@ import * as React from 'react'
 import { NextPage } from 'next'
 import Link from 'next/link'
 
-import Page from '~/components/Page'
-import PageHeader from '~/components/Layout/PageHeader'
-import PageBody from '~/components/Layout/PageBody'
+import { Page, PageHeader, PageBody } from '~/components/Layout'
 import Navigation from '~/components/Navigation'
 import Content from '~/components/Content'
 import relativeTime from '~/utils/relativeTime'
@@ -15,7 +13,7 @@ const AboutPage: NextPage = () => (
   <Page title="About">
     <Navigation />
     <Content>
-      <PageHeader>About us.</PageHeader>
+      <PageHeader title="About us." />
       <PageBody>
         <p>
           Perge porro; Igitur ne dolorem quidem. Omnes enim iucundum motum, quo sensus hilaretur. Nam, ut sint illa vendibiliora, haec
