+++ pages/index.tsx
@@ -2,9 +2,7 @@ import * as React from 'react'
 import { NextPage } from 'next'
 import Link from 'next/link'
 
-import Page from '~/components/Page'
-import PageHeader from '~/components/Layout/PageHeader'
-import PageBody from '~/components/Layout/PageBody'
+import { Page, PageHeader, PageBody } from '~/components/Layout'
 import Navigation from '~/components/Navigation'
 import Content from '~/components/Content'
 import relativeTime from '~/utils/relativeTime'
